requirements.txt
streamlit
groq
python-docx
requests